Bulk changes to improve many lexers inner workings
Based on a suspicion that most examplefiles only exercise a small part of the
lexers, I've written some code to find suspicious parts of regular
expressions, then gone back over those to fix them.  Most of these affect
whether the regex does what it looks like it does, but none of them should
appreciably change the function of the lexer.

* a few cases which used capturing groups + bygroups incorrectly (most were
  harmless, but I think one could have generated a traceback in its previous
  state)
* a few cases which could match empty string, without a callback (this is
  highly discouraged, because if an op doesn't consume any characters, it
  might be possible to enter an empty loop).  I'll revisit
  individually the cases where a callback or state push is used.
* many cases with embedded newlines in non-verbose regexes
* many, many cases with reversed (else|elseif) style alternations